+++ src/databricks/labs/ucx/assessment/azure.py
@@ -1,4 +1,3 @@
-import base64
 import json
 import re
 from collections.abc import Iterable
@@ -10,11 +9,10 @@ from databricks.sdk.service.compute import ClusterSource, Policy
 
 from databricks.labs.ucx.assessment.crawlers import azure_sp_conf_present_check, logger
 from databricks.labs.ucx.assessment.jobs import JobsMixin
+from databricks.labs.ucx.assessment.secrets import SecretsMixin
 from databricks.labs.ucx.framework.crawlers import CrawlerBase, SqlBackend
 
-SECRET_PATTERN = r"{{secrets\/(.*)\/(.*)}}"
-TENANT_PATTERN = r"https:\/\/login\.(?:partner\.)?microsoftonline\.(?:com|us|cn)\/(.*)\/oauth2\/token"
-STORAGE_ACCOUNT_EXTRACT_PATTERN = r"(?:id|endpoint)(.*?)dfs"
+_STORAGE_ACCOUNT_PATTERN = r"(?:id|endpoint)(.*?)dfs"
 
 
 @dataclass(frozen=True)
@@ -31,7 +29,7 @@ class AzureServicePrincipalInfo:
     storage_account: str | None = None
 
 
-class AzureServicePrincipalCrawler(CrawlerBase[AzureServicePrincipalInfo], JobsMixin):
+class AzureServicePrincipalCrawler(CrawlerBase[AzureServicePrincipalInfo], JobsMixin, SecretsMixin):
     def __init__(self, ws: WorkspaceClient, sbe: SqlBackend, schema):
         super().__init__(sbe, "hive_metastore", schema, "azure_service_principals", AzureServicePrincipalInfo)
         self._ws = ws
@@ -141,21 +139,21 @@ class AzureServicePrincipalCrawler(CrawlerBase[AzureServicePrincipalInfo], JobsM
             client_secret_key = "fs.azure.account.oauth2.client.secret"
 
             # retrieve storage account configured with this spn
-            storage_account_matched = re.search(STORAGE_ACCOUNT_EXTRACT_PATTERN, spn_app_id_key)
+            storage_account_matched = re.search(_STORAGE_ACCOUNT_PATTERN, spn_app_id_key)
             if storage_account_matched:
                 storage_account = storage_account_matched.group(1).strip(".")
                 # adjust the key to lookup for tenant id & secret id
-                tenant_key = tenant_key + f".{storage_account}"
-                client_secret_key = client_secret_key + f".{storage_account}"
+                tenant_key = f"{tenant_key}.{storage_account}"
+                client_secret_key = f"{client_secret_key}.{storage_account}"
 
             # retrieve client secret of spn
-            matching_secret_keys = [key for key in config.keys() if re.search(client_secret_key, key)]
+            matching_secret_keys = [key for key in config.keys() if client_secret_key in key]
             if len(matching_secret_keys) == 0 or matching_secret_keys[0] is None:
                 secret_scope, secret_key = None, None
             else:
                 secret_scope, secret_key = self._get_client_secret(config, matching_secret_keys[0])
             # retrieve tenant id of spn
-            matching_tenant_keys = [key for key in config.keys() if re.search(tenant_key, key)]
+            matching_tenant_keys = [key for key in config.keys() if tenant_key in key]
             if len(matching_tenant_keys) == 0 or matching_tenant_keys[0] is None:
                 tenant_id = None
             else:
@@ -172,52 +170,3 @@ class AzureServicePrincipalCrawler(CrawlerBase[AzureServicePrincipalInfo], JobsM
                 )
             )
         return set_service_principals
-
-    def _get_value_from_config_key(self, config: dict, key: str, get_secret: bool = True) -> str | None:
-        """Get a config value based on its key, with some special handling:
-        If the key is prefixed with spark_conf, i.e. this is in a cluster policy, the actual value is nested
-        If the value is of format {{secret_scope/secret}}, we extract that as well
-        """
-        if re.search("spark_conf", key):
-            value = config.get(key, {}).get("value", "")
-        else:
-            value = config.get(key, "")
-        # retrieve from secret scope if used
-        if not get_secret:
-            return value
-        secret_matched = re.findall(SECRET_PATTERN, value)
-        if len(secret_matched) == 0:
-            return value
-        secret_scope, secret_key = secret_matched[0][0], secret_matched[0][1]
-        return self._get_secret_if_exists(secret_scope, secret_key)
-
-    def _get_secret_if_exists(self, secret_scope, secret_key) -> str | None:
-        """Get the secret value given a secret scope & secret key. Log a warning if secret does not exist"""
-        try:
-            # Return the decoded secret value in string format
-            secret = self._ws.secrets.get_secret(secret_scope, secret_key)
-            assert secret.value is not None
-            return base64.b64decode(secret.value).decode("utf-8")
-        except NotFound:
-            logger.warning(f'removed on the backend: {secret_scope}{secret_key}')
-            return None
-
-    def _get_client_secret(self, config: dict, secret_key: str) -> tuple[str | None, str | None]:
-        client_secret = self._get_value_from_config_key(config, secret_key, False)
-        if client_secret is None:
-            return None, None
-        secret_matched = re.findall(SECRET_PATTERN, client_secret)
-        if len(secret_matched) == 0:
-            logger.warning('Secret in config stored in plaintext.')
-            return None, None
-        return secret_matched[0][0], secret_matched[0][1]
-
-    def _get_tenant_id(self, config: dict, tenant_key: str) -> str | None:
-        tenant_value = self._get_value_from_config_key(config, tenant_key)
-        if tenant_value is None:
-            return None
-        tenant_matched = re.findall(TENANT_PATTERN, tenant_value)
-        if len(tenant_matched) == 0:
-            logger.warning('Tenant id configuration is not in correct format')
-            return None
-        return tenant_matched[0]
